+++ src/app/about/components/Travelling/index.tsx
@@ -1,4 +1,4 @@
-import SectionTitle from 'src/components/shared/SectionTitle'
+import SectionTitle from 'src/components/SectionTitle'
 import Label from './Label'
 import Map from './Map'
