+++ src/components/Layout/NavigationMenu/Hamburger/CloseIcon/index.tsx
