+++ src/components/Layout/NavigationMenu/Hamburger/OpenIcon/index.tsx
