+++ src/components/Layout/NavigationMenu/ThemeSelector/Icon/index.tsx
