+++ src/components/README.md
@@ -3,6 +3,6 @@
 This directory contains the components that **will be shared by the whole application**.
 
 🖐 **Components** that are used on a single page, should be placed into his
-specific folder under the `pages/` directory.
+specific folder under the `app/` directory.
 
-Components that are used on multiple places should be placed into the `shared/` directory.
+Components that are used on multiple places should be placed here.
