+++ src/components/SectionTitle/__tests__/sectionTitle.spec.tsx
