+++ src/components/SectionTitle/__tests__/stubs.ts
