+++ src/components/SectionTitle/index.tsx
